chore: breaking apart elements.ts into separate files by concern remove circular dependencies
